♻️ refactor: Page component refactor. home -> map / login -> default page.
